Meta information for content listings
Modifies the DB storage of content items to be of a hash type instead of a simple string type, providing the capability to store meta information about content files.

parseContent() has been added, which reads the first few lines of a content file for meta information in the following format:

[key] value

As soon as a line has been reached which does not meet this format (including a blank line), the rest of the document will be treated as the markdown body.

Templates have been updated appropriately.

Oh, and the application port has been changed to 8000.